# portal/src/static/openapi.yml
openapi: 3.0.3
info:
  title: Healthcare Services API
  description: This is our Healthcare Services API that let you develop healthcare applications effortlessly.
  version: "1.0"
servers:
  - url: http://localhost:8002
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/er:
    get:
      tags:
        - "ER Record Management"
      description: Get all ER records
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                [{"patient_id": "665186afb873a548422f9b79",
                  "err": [{"ID": "20f0c670-c37a-4ba6-8a46-ff28e01a9502",
                            "chief_complaint": "Head Injury",
                            "date": "2024-05-20"}
                            ]},
                  {"patient_id": "665186dbb873a548422f9b7b",
                    "err": [{"ID": "8af925de-8d15-458f-9f25-ea85104760f5",
                              "date": "2024-05-25",
                              "chief_complaint": "Broken Nose"}]}]
  /api/er/byDate:
    get:
      tags:
        - "ER Record Management"
      description: Get ER Records by Date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            example: "2024-06-04"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                {"ER Records": [{"patient_id": "665186afb873a548422f9b79",
                                  "err": [{"ID": "dc28f399-9f8b-4200-ab3c-9fd107883b59",
                                            "date": "2024-06-04",
                                            "chief_complaint": "Head Injury"},
                                          ]},
                                {"patient_id": "665186dbb873a548422f9b7b",
                                "err": [{"ID": "8af925de-8d15-458f-9f25-ea85104760f5",
                                          "date": "2024-06-04",
                                          "chief_complaint": "Broken Nose"}]}]}
  /api/er/{patient_id}:
    get:
      tags:
        - "ER Record Management"
      description: Get ER records of a particular Patient
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: 665186afb873a548422f9b79
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                {"ER Records": [{"ID": "20f0c670-c37a-4ba6-8a46-ff28e01a9502",
                                  "date": "2024-05-20",
                                  "chief_complaint": "Head Injury"},
                                  {"ID": "90449f92-ae2a-4de4-80f6-48b90b707f57",
                                    "date": "2024-03-12",
                                    "chief_complaint": "Broken ribs"}]}

    post:
      tags:
        - "ER Record Management"
      description: Add ER Record
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: 665186afb873a548422f9b79
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                doctor:
                  type: string
                  example: <doctor's name here>
                date:
                  type: string
                  example: <date here>
                arrival_time:
                  type: string
                  example: <arrival_time here>
                chief_complaint:
                  type: string
                  example: <chief_complaint here>
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                message: "Record added Successfully!"
  /api/er/{patient_id}/{er_id}:
    get:
      tags:
        - "ER Record Management"
      description: Get a specific ER Record
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: "665186afb873a548422f9b79"
        - name: er_id
          in: path
          required: true
          schema:
            type: string
            example: "20f0c670-c37a-4ba6-8a46-ff28e01a9502"
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                { "ER Record": {"id": "20f0c670-c37a-4ba6-8a46-ff28e01a9502",
                                "chief_complaint": "Head Injury",
                                "date": "2024-05-27",
                                "arrival_time": "10:35:42",
                                "doctor": "Doctor 1"
                                }}

    delete:
      tags:
        - "ER Record Management"
      description: Delete a specific ER Record
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: "665186afb873a548422f9b79"
        - name: er_id
          in: path
          required: true
          schema:
            type: string
            example: "20f0c670-c37a-4ba6-8a46-ff28e01a9502"
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                message: Record Deleted Successfully!
  /api/er/byDate/{patient_id}:
    get:
      tags:
        - "ER Record Management"
      description: Get a specific Patient's ER Records by Date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            example: "2024-06-04"
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
            example: "665186afb873a548422f9b79"
      responses:
        200:
          description: Success
          content:
            application/json:
              example:
                [{"ER Records": [{"ID": "dc28f399-9f8b-4200-ab3c-9fd107883b59",
                                    "date": "2024-06-04",
                                    "chief_complaint": "Head Injury"
                                    }
                                  ]}]
...